+++ sedaro/src/sedaro/utils.py
@@ -14,6 +14,17 @@ from .settings import BLOCKS, COMMON_API_KWARGS, STATUS
 if TYPE_CHECKING:
     from .branches.branch import Branch
 
+def serdes(v):
+    import numpy as np
+    if type(v) is dict and 'ndarray' in v:
+        return np.array(v['ndarray'])
+    if type(v) is np.ndarray:
+        return {'ndarray': v.tolist()}
+    if type(v) is dict:
+        return {k: serdes(v) for k, v in v.items()}
+    if type(v) in {list, tuple}:
+        return [serdes(v) for v in v]
+    return v
 
 def parse_urllib_response(response: HTTPResponse) -> Dict:
     '''Parses the response from urllib3.response.HTTPResponse into a dictionary'''
